Staging (#23)
* Exporting alleles of INDELs with unknown sequence as I/D

for GFF and Hapmap formats

* Export optim (#12)

* Modified Hapmap export to be multithreaded

* Parallelized export working for all variant-oriented formats

* Minor change

* Fixed issues with writing variant-list files when exporting

* Modified Hapmap export to make it able to write IGV data

* Individual-oriented exports now working in parallel

* Moved class from Mgdb2 project

* Accounting for assembly to read positions in individual-oriented exports

* Re-added header to top of SNP file

* Minor change

* Added mean to specify a location for export extraction files

* All seems to be working fine

* Ensured temporary warning file gets deleted

* Fixed missing allele list bug that was happening in some cases

... that is, when some alleles were unknown to the VRD record

* Merge pull request #11 from GuilhemSempere/staging (#15)

* POM version update

* Minor change

* JK tree export is no more an ExperimentalFeature

* Added means to dump distance matrix to file (commented out for now)

* Turned NaN values into 3 (bugfix)

* Moved VisualizationService class from Gigwa2ServiceImpl to Mgdb2Export

* Made readToken method static

* Some refactoring

* Moving towards supporting visualization on multiple projects

* Working towards making VizService handle multiple projects' data

* Made charts able to account for several projects at once

* Working around exceptions thrown because of export aborts

* Harmonized the way we build export queries

* Moving towards making UI & exports support working with samples

* Made VizService compatible with detached samples

* Minor fix

* Made IGV data export compatible with detached samples

* Minor change

* Import refactoring and Split Sample into 2 new collections : IndSample and CallSets

* Remove pj and rn fields from GenotypingSample and update queries to get callsets from individuals

* Minor fix

* Minor change

* Minor changes

* Further integrating multi-projet UI + callSet level addition

* Minor fix

* CallSets now embedded in Samples

* Fixing errors due to addition of CallSet level

* Ready for testing

* Fixed handling of genotypes for multi-callset individuals

* Minor change

* Minor change

* Removed commented-out code

* Minor fix

* Fixed remaining bugs in accounting for callSets in computing graph data

* Optimized Jukes-Cantor distance calculation procedure

* Added export-handler for Jukes-cantor distance matrix

* Callsets multi proj (#21)

* Moving towards supporting visualization on multiple projects

* Working towards making VizService handle multiple projects' data

* Made charts able to account for several projects at once

* Moving towards making UI & exports support working with samples

* Made VizService compatible with detached samples

* Minor fix

* Made IGV data export compatible with detached samples

* Minor change

* Import refactoring and Split Sample into 2 new collections : IndSample and CallSets

* Remove pj and rn fields from GenotypingSample and update queries to get callsets from individuals

* Minor fix

* Minor change

* Minor changes

* Further integrating multi-projet UI + callSet level addition

* Minor fix

* CallSets now embedded in Samples

* Fixing errors due to addition of CallSet level

* Ready for testing

* Fixed handling of genotypes for multi-callset individuals

* Minor change

* Minor change

* Removed commented-out code

* Minor fix

* Fixed remaining bugs in accounting for callSets in computing graph data

* Optimized Jukes-Cantor distance calculation procedure

* Added export-handler for Jukes-cantor distance matrix

---------


* Pca export (#22)

* Changed missing data encoding from N to -

* Implemented NJ tree export

* Fixed package name, removed unused classes

* Fixed package name, removed unused classes

* Started implementation of PCA export

* Attempting to use less RAM

* Attempting to reduce required RAM (unfinished)

* Fixed conflicts + reduced memory usage

* Switched back on matrix size limitation for PCA calculation

* Fixed missing allele list bug that was happening in some cases (#13)

... that is, when some alleles were unknown to the VRD record

* Fixed missing allele list bug that was happening in some cases (#14)

... that is, when some alleles were unknown to the VRD record

* Merge from staging (#16)

* Fixed missing allele list bug that was happening in some cases

... that is, when some alleles were unknown to the VRD record

* Merge pull request #11 from GuilhemSempere/staging (#15)

* Merge from staging (#17)

* Fixed missing allele list bug that was happening in some cases

... that is, when some alleles were unknown to the VRD record

* Merge pull request #11 from GuilhemSempere/staging (#15)

* POM version update

* Staging (#18)

* Fixed missing allele list bug that was happening in some cases

... that is, when some alleles were unknown to the VRD record

* Merge pull request #11 from GuilhemSempere/staging (#15)

* POM version update

* Significant improvement in lowering memory usage

(computation is still not parallelized though)

* Made sure memory-mapped-files are deleted if an error occurs

* Staging (#19)

* Fixed missing allele list bug that was happening in some cases

... that is, when some alleles were unknown to the VRD record

* Merge pull request #11 from GuilhemSempere/staging (#15)

* POM version update

* Minor change

* JK tree export is no more an ExperimentalFeature

* Made SVD calculation abortable, set numoutevec to 20

* PCA calculation now able to cope with larges matrices (via segmentation)

* Improved generated VCF header generation when exporting merged VCF runs

---------

+++ src/fr/cirad/mgdb/exporting/individualoriented/PLinkExportHandler.java
@@ -18,6 +18,7 @@ package fr.cirad.mgdb.exporting.individualoriented;
 
 import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileReader;
 import java.io.IOException;
@@ -41,7 +42,6 @@ import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
 
 import com.mongodb.BasicDBList;
-import com.mongodb.BasicDBObject;
 import com.mongodb.client.MongoCursor;
 
 import fr.cirad.mgdb.exporting.IExportHandler;
@@ -109,7 +109,7 @@ public class PLinkExportHandler extends AbstractIndividualOrientedExportHandler 
 	}
 
     @Override
-    public void exportData(OutputStream outputStream, String sModule, Integer nAssemblyId, String sExportingUser, ExportOutputs exportOutputs, boolean fDeleteSampleExportFilesOnExit, ProgressIndicator progress, String tmpVarCollName, VariantQueryWrapper varQueryWrapper, long markerCount, Map<String, String> markerSynonyms, Collection<String> individualMetadataFieldsToExport, Map<String, String> individualPopulations, Map<String, InputStream> readyToExportFiles) throws Exception {
+    public void exportData(OutputStream outputStream, String sModule, Integer nAssemblyId, ExportOutputs exportOutputs, boolean fDeleteSampleExportFilesOnExit, ProgressIndicator progress, String tmpVarCollName, VariantQueryWrapper varQueryWrapper, long markerCount, Map<String, String> markerSynonyms, Map<String, String> individualPopulations, Map<String, InputStream> readyToExportFiles) throws Exception {
 		MongoTemplate mongoTemplate = MongoTemplateManager.get(sModule);
         int nQueryChunkSize = IExportHandler.computeQueryChunkSize(mongoTemplate, markerCount);
         
@@ -128,22 +128,24 @@ public class PLinkExportHandler extends AbstractIndividualOrientedExportHandler 
 		    	}
 	        }
 	
-	        ZipOutputStream zos = IExportHandler.createArchiveOutputStream(outputStream, readyToExportFiles);
+	        ZipOutputStream zos = IExportHandler.createArchiveOutputStream(outputStream, readyToExportFiles, exportOutputs);
+
 			Assembly assembly = mongoTemplate.findOne(new Query(Criteria.where("_id").is(nAssemblyId)), Assembly.class);
-	        String exportName = sModule + (assembly != null && assembly.getName() != null ? "__" + assembly.getName() : "") + "__" + markerCount + "variants__" + exportOutputs.getGenotypeFiles().length + "individuals";
+			String exportName = IExportHandler.buildExportName(sModule, assembly, markerCount, exportOutputs.getGenotypeFiles().length, exportOutputs.isWorkWithSamples());
 	        
 	        ArrayList<String> exportedIndividuals = new ArrayList<>();
 	        for (File indFile : exportOutputs.getGenotypeFiles())
 	        	try (Scanner scanner = new Scanner(indFile)) {
 	        		exportedIndividuals.add(scanner.nextLine());
 	        	}
-	
-	        if (individualMetadataFieldsToExport == null || !individualMetadataFieldsToExport.isEmpty())
-	        	IExportHandler.addMetadataEntryIfAny(sModule + "__" + exportOutputs.getGenotypeFiles().length + "individuals_metadata.tsv", sModule, sExportingUser, exportedIndividuals, individualMetadataFieldsToExport, zos, "individual");
+		        catch (Exception e) {
+		        	if (!(e instanceof NullPointerException || e instanceof IOException || e instanceof FileNotFoundException) || !progress.isAborted())
+		        		throw e;
+		        }
 	
 	        Collection<BasicDBList> variantDataQueries = varQueryWrapper.getVariantDataQueries();
-	        BasicDBObject varQuery = !variantDataQueries.isEmpty() ? new BasicDBObject("$and", variantDataQueries.iterator().next()) : new BasicDBObject();
-	
+	        Document variantQueryForTargetCollection = variantDataQueries.isEmpty() ? new Document() : (tmpVarCollName == null ? new Document("$and", variantDataQueries.iterator().next()) : (varQueryWrapper.getBareQueries().iterator().hasNext() ? new Document("$and", varQueryWrapper.getBareQueries().iterator().next()) : new Document()));
+	        
 	        zos.putNextEntry(new ZipEntry(exportName + ".ped"));
 	        writeGenotypeFile(zos, sModule, exportedIndividuals, individualPopulations, nQueryChunkSize, markerSynonyms, exportOutputs.getGenotypeFiles(), warningOS, progress);
 	    	zos.closeEntry();
@@ -154,7 +156,7 @@ public class PLinkExportHandler extends AbstractIndividualOrientedExportHandler 
 	        ArrayList<Comparable> unassignedMarkers = new ArrayList<>();
 	    	String refPosPathWithTrailingDot = Assembly.getThreadBoundVariantRefPosPath() + ".";
 	    	Document projectionAndSortDoc = new Document(refPosPathWithTrailingDot + ReferencePosition.FIELDNAME_SEQUENCE, 1).append(refPosPathWithTrailingDot + ReferencePosition.FIELDNAME_START_SITE, 1);
-			try (MongoCursor<Document> markerCursor = IExportHandler.getMarkerCursorWithCorrectCollation(mongoTemplate.getCollection(tmpVarCollName != null ? tmpVarCollName : mongoTemplate.getCollectionName(VariantData.class)), tmpVarCollName != null ? new Document() : new Document(varQuery), projectionAndSortDoc, nQueryChunkSize)) {
+			try (MongoCursor<Document> markerCursor = IExportHandler.getMarkerCursorWithCorrectCollation(mongoTemplate.getCollection(tmpVarCollName != null ? tmpVarCollName : mongoTemplate.getCollectionName(VariantData.class)), variantQueryForTargetCollection, projectionAndSortDoc, nQueryChunkSize)) {
 	            progress.addStep("Writing map file");
 	            progress.moveToNextStep();
 		        while (markerCursor.hasNext()) {
@@ -299,9 +301,10 @@ public class PLinkExportHandler extends AbstractIndividualOrientedExportHandler 
         finally
         {
         	for (File f : individualExportFiles)
-                if (!f.delete()) {
+                if (f != null && !f.delete()) {
                     f.deleteOnExit();
-                    LOG.info("Unable to delete tmp export file " + f.getAbsolutePath());
+                    if (!progress.isAborted())
+                    	LOG.info("Unable to delete tmp export file " + f.getAbsolutePath());
                 }
         }
     }
